+++ lib/atomic-ruby/atomic_count_down_latch.rb
@@ -13,6 +13,8 @@ module AtomicRuby
       end
 
       @count = Atom.new(count)
+
+      Ractor.make_shareable(self)
     end
 
     def count